requirements.txt
argcomplete==1.12.3
asgiref==3.3.4
awscli==1.19.68
boto3==1.17.67
botocore==1.20.68
certifi==2020.12.5
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
colorama==0.4.3
Django==3.2.2
djangorestframework==3.12.4
docutils==0.15.2
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==2.10
jmespath==0.10.0
kappa==0.6.0
pep517==0.10.0
pip-tools==6.1.0
placebo==0.9.0
pyasn1==0.4.8
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.60.0
troposphere==2.7.1
urllib3==1.26.4
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
